Unify sickdays into sicknotes/sickdays
