+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/statistics/SickNoteStatisticsService.java
@@ -5,9 +5,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.synyx.urlaubsverwaltung.department.DepartmentService;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.person.PersonId;
-import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedata;
-import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedataService;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNote;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNoteService;
 import org.synyx.urlaubsverwaltung.workingtime.WorkDaysCountService;
@@ -15,14 +12,10 @@ import org.synyx.urlaubsverwaltung.workingtime.WorkDaysCountService;
 import java.time.Clock;
 import java.time.LocalDate;
 import java.time.Year;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
 import java.util.stream.Stream;
 
 import static java.time.temporal.TemporalAdjusters.lastDayOfYear;
-import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
 import static org.synyx.urlaubsverwaltung.person.Role.BOSS;
 import static org.synyx.urlaubsverwaltung.person.Role.DEPARTMENT_HEAD;
@@ -41,14 +34,12 @@ public class SickNoteStatisticsService {
     private final SickNoteService sickNoteService;
     private final WorkDaysCountService workDaysCountService;
     private final DepartmentService departmentService;
-    private final PersonBasedataService personBasedataService;
 
     @Autowired
-    SickNoteStatisticsService(SickNoteService sickNoteService, WorkDaysCountService workDaysCountService, DepartmentService departmentService, PersonBasedataService personBasedataService) {
+    SickNoteStatisticsService(SickNoteService sickNoteService, WorkDaysCountService workDaysCountService, DepartmentService departmentService) {
         this.sickNoteService = sickNoteService;
         this.workDaysCountService = workDaysCountService;
         this.departmentService = departmentService;
-        this.personBasedataService = personBasedataService;
     }
 
     SickNoteStatistics createStatisticsForPerson(Person person, Clock clock) {
@@ -57,42 +48,6 @@ public class SickNoteStatisticsService {
         final LocalDate lastDayOfYear = firstDayOfYear.with(lastDayOfYear());
         final List<SickNote> sickNotes = getSickNotes(person, firstDayOfYear, lastDayOfYear);
         return new SickNoteStatistics(clock, sickNotes, workDaysCountService);
-    }
-
-    /**
-     * Returns a list of all sick notes detailed statistics that the person is allowed to access.
-     *
-     * @param person to ask for the statistics
-     * @param from   a specific date
-     * @param to     a specific date
-     * @return list of all {@link SickNoteDetailedStatistics} that the person can access
-     */
-    List<SickNoteDetailedStatistics> getAll(Person person, LocalDate from, LocalDate to) {
-
-        final List<SickNote> sickNotes = getSickNotes(person, from, to);
-        final Map<Person, List<SickNote>> sickNotesByPerson = sickNotes.stream()
-            .collect(groupingBy(SickNote::getPerson));
-
-        final List<Person> personsWithSickNotes = new ArrayList<>(sickNotesByPerson.keySet());
-
-        final List<Integer> personIds = personsWithSickNotes.stream().map(Person::getId).collect(toList());
-        final Map<PersonId, PersonBasedata> basedataForPersons = personBasedataService.getBasedataByPersonId(personIds);
-        final Map<PersonId, List<String>> departmentsForPersons = departmentService.getDepartmentNamesByMembers(personsWithSickNotes);
-
-        return sickNotesByPerson.entrySet().stream()
-            .map(toSickNoteDetailedStatistics(basedataForPersons, departmentsForPersons))
-            .collect(toList());
-    }
-
-    private Function<Map.Entry<Person, List<SickNote>>, SickNoteDetailedStatistics> toSickNoteDetailedStatistics(Map<PersonId, PersonBasedata> basedataForPersons, Map<PersonId, List<String>> departmentsForPersons) {
-        return personListEntry ->
-        {
-            final Person person = personListEntry.getKey();
-            final PersonId personId = new PersonId(person.getId());
-            final String personnelNumber = basedataForPersons.getOrDefault(personId, new PersonBasedata(personId, "", "")).getPersonnelNumber();
-            final List<String> departments = departmentsForPersons.getOrDefault(personId, List.of());
-            return new SickNoteDetailedStatistics(personnelNumber, person.getFirstName(), person.getLastName(), personListEntry.getValue(), departments);
-        };
     }
 
     private List<SickNote> getSickNotes(Person person, LocalDate from, LocalDate to) {
