+++ src/test/java/org/synyx/urlaubsverwaltung/overview/SickDaysOverviewTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.synyx.urlaubsverwaltung.TestDataCreator;
 import org.synyx.urlaubsverwaltung.period.DayLength;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.sickdays.web.SickDays;
+import org.synyx.urlaubsverwaltung.sicknote.sickdays.SickDays;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNote;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNoteCategory;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNoteStatus;
